refactor p2
