+++ Cargo.toml
@@ -11,9 +11,7 @@ members = [
     "ddtelemetry-ffi",
     "tools",
     "ipc",
-    "ipc/macros",
     "tools/cc_utils",
-    "tools/proc_macros",
     "trace-normalization",
     "trace-obfuscation",
     "spawn_worker",
