+++ ipc/tests/flock.rs
@@ -8,7 +8,7 @@ use std::{
     time::Duration,
 };
 
-use ddtelemetry::ipc::platform::locks::FLock;
+use datadog_ipc::platform::locks::FLock;
 use spawn_worker::{assert_child_exit, entrypoint, fork::set_default_child_panic_handler, Stdio};
 use tempfile::tempdir;
