--- tools/proc_macros/Cargo.toml
@@ -1,12 +1,0 @@
-[package]
-name = "dd-proc-macros"
-version = "0.1.0"
-edition = "2021"
-author = "P"
-
-[lib]
-proc-macro = true
-
-[dependencies]
-syn = "1.0"
-quote = "1.0"
